Добавил какую-то часть я жуе заеб

+++ backend/FunApi/Models/Carts/Cart.cs
@@ -4,8 +4,8 @@ namespace FunApi.Models.Carts
 {
     public class Cart
     {
-        public long Id { get; set; }
-        public long UserId { get; set; }
+        public int Id { get; set; }
+        public int UserId { get; set; }
 
         public DateTime CreatedAt { get; set; }
         public DateTime UpdatedAt { get; set; }
